Refactor into views, components, elements
