+++ src/elements/Content.jsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+import tw from 'tailwind.macro'
+import { ParallaxLayer } from 'react-spring/addons.cjs'
+
+const Content = styled(ParallaxLayer)`
+  ${tw`p-6 md:p-12 lg:p-24 justify-center items-center flex z-50`};
+`
+
+export default Content
