+++ src/pages/index.jsx
@@ -1,65 +1,23 @@
 import React from 'react'
 import styled from 'styled-components'
 import tw from 'tailwind.macro'
-import { Parallax, ParallaxLayer } from 'react-spring/addons.cjs'
+import { Parallax } from 'react-spring/addons.cjs'
+
+// Components
 import Layout from '../components/Layout'
-import SVG from '../components/SVG'
 import ProjectCard from '../components/ProjectCard'
-import { rotateAnimation, UpDown, UpDownWide, waveAnimation } from '../styles/animations'
-import { colors } from '../../tailwind'
-import triangle from '../images/triangle.svg'
+
+// Elements
+import Inner from '../elements/Inner'
+import { Title, BigTitle, Subtitle } from '../elements/Titles'
+
+// Views
+import Hero from '../views/Hero'
+import Projects from '../views/Projects'
+import About from '../views/About'
+import Contact from '../views/Contact'
+
 import avatar from '../images/avatar.jpg'
-
-const Divider = styled(ParallaxLayer)`
-  ${tw`absolute w-full h-full`};
-  background: ${props => props.bg};
-  svg {
-    fill: ${props => props.fill};
-  }
-  clip-path: ${props => props.clipPath};
-`
-
-const DividerMiddle = styled(Divider)`
-  clip-path: polygon(0 15%, 100% 25%, 100% 85%, 0 75%);
-`
-
-const Content = styled(ParallaxLayer)`
-  ${tw`p-6 md:p-12 lg:p-24 justify-center items-center flex z-50`};
-`
-
-const Hero = styled.div`
-  ${tw`w-full xl:w-2/3`};
-`
-
-const Inner = styled.div`
-  ${tw`w-full xxl:w-2/3 text-center lg:text-left`};
-`
-
-const BigTitle = styled.h1`
-  ${tw`text-5xl lg:text-6xl font-serif text-white mb-6 tracking-wide`};
-  text-shadow: 0 5px 35px rgba(255, 255, 255, 0.15);
-`
-
-const Title = styled.h1`
-  ${tw`text-4xl lg:text-4xl font-serif text-white mb-8 tracking-wide relative inline-block`};
-  text-shadow: 0 2px 10px rgba(0, 0, 0, 0.2);
-  &:before {
-    content: '';
-    width: 40px;
-    height: 40px;
-    background: url(${triangle});
-    position: absolute;
-    background-size: 40px;
-    ${rotateAnimation('4s')};
-    left: -60px;
-    top: 5px;
-  }
-`
-
-const Subtitle = styled.p`
-  ${tw`text-2xl lg:text-4xl font-sans text-white mt-8 xxl:w-3/4`};
-  text-shadow: 0 2px 15px rgba(0, 0, 0, 0.2);
-`
 
 const ProjectsWrapper = styled.div`
   ${tw`flex flex-wrap justify-between mt-8`};
@@ -72,22 +30,6 @@ const ProjectsWrapper = styled.div`
   @media (max-width: 900px) {
     grid-template-columns: 1fr;
     grid-gap: 2rem;
-  }
-`
-
-const WaveWrapper = styled.div`
-  ${tw`absolute pin-b w-full`};
-  transform: matrix(1, 0, 0, -1, 0, 0);
-`
-
-const InnerWave = styled.div`
-  ${tw`relative h-full`};
-  svg {
-    width: 100%;
-    height: 40vh;
-  }
-  path {
-    ${waveAnimation('20s')};
   }
 `
 
@@ -119,160 +61,63 @@ const Index = () => (
   <>
     <Layout />
     <Parallax pages={5}>
-      <Divider speed={0.2} offset={0}>
-        <UpDown>
-          <SVG icon="triangle" hiddenMobile width={48} stroke={colors.orange} left="10%" top="20%" />
-          <SVG icon="hexa" width={48} stroke={colors.red} left="60%" top="70%" />
-          <SVG icon="box" width={6} fill={colors['grey-darker']} left="60%" top="15%" />
-        </UpDown>
-        <UpDownWide>
-          <SVG icon="arrowUp" hiddenMobile width={16} fill={colors['blue-dark']} left="80%" top="10%" />
-          <SVG icon="triangle" width={12} stroke={colors.white} left="90%" top="50%" />
-          <SVG icon="circle" width={16} fill={colors['grey-darker']} left="70%" top="90%" />
-          <SVG icon="triangle" width={16} stroke={colors['grey-darkest']} left="30%" top="65%" />
-          <SVG icon="circle" width={6} fill={colors['grey-darkest']} left="75%" top="10%" />
-          <SVG icon="upDown" hiddenMobile width={8} fill={colors['grey-darkest']} left="45%" top="10%" />
-        </UpDownWide>
-        <SVG icon="circle" hiddenMobile width={24} fill={colors['grey-darker']} left="5%" top="70%" />
-        <SVG icon="circle" width={6} fill={colors['grey-darkest']} left="4%" top="20%" />
-        <SVG icon="circle" width={12} fill={colors['grey-darkest']} left="50%" top="60%" />
-        <SVG icon="upDown" width={8} fill={colors['grey-darkest']} left="95%" top="90%" />
-        <SVG icon="upDown" hiddenMobile width={24} fill={colors['grey-darker']} left="40%" top="80%" />
-        <SVG icon="triangle" width={8} stroke={colors['grey-darker']} left="25%" top="5%" />
-        <SVG icon="circle" width={64} fill={colors.green} left="95%" top="5%" />
-        <SVG icon="box" hiddenMobile width={64} fill={colors.purple} left="5%" top="90%" />
-        <SVG icon="box" width={6} fill={colors['grey-darkest']} left="10%" top="10%" />
-        <SVG icon="box" width={12} fill={colors['grey-darkest']} left="40%" top="30%" />
-        <SVG icon="hexa" width={16} stroke={colors['grey-darker']} left="10%" top="50%" />
-        <SVG icon="hexa" width={8} stroke={colors['grey-darker']} left="80%" top="70%" />
-      </Divider>
-      <Content speed={0.4} offset={0}>
-        <Hero>
-          <BigTitle>
-            Hello, <br /> I'm John Doe.
-          </BigTitle>
-          <Subtitle>I'm creating noice web experiences for the next generation of consumer-facing companies.</Subtitle>
-        </Hero>
-      </Content>
-      <DividerMiddle
-        bg="linear-gradient(to right, SlateBlue 0%, DeepSkyBlue 100%)"
-        speed={-0.2}
-        offset={1.1}
-        factor={2}
-      />
-      <Content speed={0.4} offset={1.2} factor={2}>
-        <Inner>
-          <Title>Projects</Title>
-          <ProjectsWrapper>
-            <ProjectCard
-              title="Freiheit"
-              link="https://www.behance.net/gallery/58937147/Freiheit"
-              bg="linear-gradient(to right, #D4145A 0%, #FBB03B 100%)"
-            >
-              This project is my entry to Adobe's #ChallengeYourPerspective contest.
-            </ProjectCard>
-            <ProjectCard
-              title="Harry Potter"
-              link="https://www.behance.net/gallery/52915793/Harry-Potter"
-              bg="linear-gradient(to right, #662D8C 0%, #ED1E79 100%)"
-            >
-              I entered the DOCMA 2017 award with this Harry Potter inspired image.
-            </ProjectCard>
-            <ProjectCard
-              title="Tomb Raider"
-              link="https://www.behance.net/gallery/43907099/Tomb-Raider"
-              bg="linear-gradient(to right, #009245 0%, #FCEE21 100%)"
-            >
-              Recreation of a Tomb Raider Wallpaper (Fan Art)
-            </ProjectCard>
-            <ProjectCard
-              title="Eagle"
-              link="https://www.behance.net/gallery/38068151/Eagle"
-              bg="linear-gradient(to right, #D585FF 0%, #00FFEE 100%)"
-            >
-              A fantasy image manipulation relocating the habitat of wild animals.
-            </ProjectCard>
-          </ProjectsWrapper>
-        </Inner>
-      </Content>
-      <Divider speed={0.1} offset={1} factor={2}>
-        <UpDown>
-          <SVG icon="box" width={6} fill={colors.white} left="85%" top="75%" />
-          <SVG icon="upDown" width={8} fill={colors.teal} left="70%" top="20%" />
-          <SVG icon="triangle" width={8} stroke={colors.orange} left="25%" top="5%" />
-          <SVG icon="circle" hiddenMobile width={24} fill={colors.white} left="17%" top="60%" />
-        </UpDown>
-        <UpDownWide>
-          <SVG icon="arrowUp" hiddenMobile width={16} fill={colors.green} left="20%" top="90%" />
-          <SVG icon="triangle" width={12} stroke={colors.white} left="90%" top="30%" />
-          <SVG icon="circle" width={16} fill={colors.yellow} left="70%" top="90%" />
-          <SVG icon="triangle" hiddenMobile width={16} stroke={colors.teal} left="18%" top="75%" />
-          <SVG icon="circle" width={6} fill={colors.white} left="75%" top="10%" />
-          <SVG icon="upDown" hiddenMobile width={8} fill={colors.green} left="45%" top="10%" />
-        </UpDownWide>
-        <SVG icon="circle" width={6} fill={colors.white} left="4%" top="20%" />
-        <SVG icon="circle" width={12} fill={colors.pink} left="80%" top="60%" />
-        <SVG icon="box" width={6} fill={colors.orange} left="10%" top="10%" />
-        <SVG icon="box" width={12} fill={colors.yellow} left="29%" top="26%" />
-        <SVG icon="hexa" width={16} stroke={colors.red} left="75%" top="30%" />
-        <SVG icon="hexa" width={8} stroke={colors.yellow} left="80%" top="70%" />
-      </Divider>
-      <Divider bg="#23262b" clipPath="polygon(0 16%, 100% 4%, 100% 82%, 0 94%)" speed={0.2} offset={3} />
-      <Divider speed={0.1} offset={3}>
-        <UpDown>
-          <SVG icon="box" hiddenMobile width={6} fill={colors.blue} left="50%" top="75%" />
-          <SVG icon="upDown" hiddenMobile width={8} fill={colors['grey-darkest']} left="70%" top="20%" />
-          <SVG icon="triangle" width={8} stroke={colors['grey-darkest']} left="25%" top="5%" />
-          <SVG icon="upDown" hiddenMobile width={24} fill={colors.orange} left="80%" top="80%" />
-        </UpDown>
-        <UpDownWide>
-          <SVG icon="arrowUp" hiddenMobile width={16} fill={colors.purple} left="5%" top="80%" />
-          <SVG icon="triangle" width={12} stroke={colors.white} left="95%" top="50%" />
-          <SVG icon="circle" width={6} fill={colors.white} left="85%" top="15%" />
-          <SVG icon="upDown" hiddenMobile width={8} fill={colors['grey-darkest']} left="45%" top="10%" />
-        </UpDownWide>
-        <SVG icon="circle" width={6} fill={colors.white} left="4%" top="20%" />
-        <SVG icon="circle" width={12} fill={colors['grey-darkest']} left="70%" top="60%" />
-        <SVG icon="box" width={6} fill={colors.orange} left="10%" top="10%" />
-        <SVG icon="box" width={12} fill={colors['grey-darkest']} left="20%" top="30%" />
-        <SVG icon="hexa" width={8} stroke={colors['grey-darkest']} left="80%" top="70%" />
-      </Divider>
-      <Content speed={0.4} offset={3}>
-        <Inner>
-          <Title>About</Title>
-          <AboutHero>
-            <Avatar src={avatar} alt="John Doe" />
-            <AboutSub>
-              The English language can not fully capture the depth and complexity of my thoughts. So I'm incorporating
-              Emoji into my speech to better express myself. Winky face.
-            </AboutSub>
-          </AboutHero>
-          <AboutDesc>
-            You know the way you feel when you see a picture of two otters holding hands? That's how you're gonna feel
-            every day. My mother cried the day I was born because she knew she’d never be prettier than me. You should
-            make me your campaign manager. I was born for politics. I have great hair and I love lying. Captain? The
-            kids want to know where Paulie the Pigeon is. I told them he got sucked up into an airplane engine, is that
-            all right?
-          </AboutDesc>
-        </Inner>
-      </Content>
-      <Divider fill="#23262b" speed={0.2} offset={4}>
-        <WaveWrapper>
-          <InnerWave>
-            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 800 338.05" preserveAspectRatio="none">
-              <path>
-                <animate
-                  attributeName="d"
-                  values="M 0 100 Q 250 50 400 200 Q 550 350 800 300 L 800 0 L 0 0 L 0 100 Z;M 0 100 Q 200 150 400 200 Q 600 250 800 300 L 800 0 L 0 0 L 0 100 Z;M 0 100 Q 150 350 400 200 Q 650 50 800 300 L 800 0 L 0 0 L 0 100 Z"
-                  repeatCount="indefinite"
-                  dur="30s"
-                />
-              </path>
-            </svg>
-          </InnerWave>
-        </WaveWrapper>
-      </Divider>
-      <Content speed={0.4} offset={4}>
+      <Hero>
+        <BigTitle>
+          Hello, <br /> I'm John Doe.
+        </BigTitle>
+        <Subtitle>I'm creating noice web experiences for the next generation of consumer-facing companies.</Subtitle>
+      </Hero>
+      <Projects>
+        <Title>Projects</Title>
+        <ProjectsWrapper>
+          <ProjectCard
+            title="Freiheit"
+            link="https://www.behance.net/gallery/58937147/Freiheit"
+            bg="linear-gradient(to right, #D4145A 0%, #FBB03B 100%)"
+          >
+            This project is my entry to Adobe's #ChallengeYourPerspective contest.
+          </ProjectCard>
+          <ProjectCard
+            title="Harry Potter"
+            link="https://www.behance.net/gallery/52915793/Harry-Potter"
+            bg="linear-gradient(to right, #662D8C 0%, #ED1E79 100%)"
+          >
+            I entered the DOCMA 2017 award with this Harry Potter inspired image.
+          </ProjectCard>
+          <ProjectCard
+            title="Tomb Raider"
+            link="https://www.behance.net/gallery/43907099/Tomb-Raider"
+            bg="linear-gradient(to right, #009245 0%, #FCEE21 100%)"
+          >
+            Recreation of a Tomb Raider Wallpaper (Fan Art)
+          </ProjectCard>
+          <ProjectCard
+            title="Eagle"
+            link="https://www.behance.net/gallery/38068151/Eagle"
+            bg="linear-gradient(to right, #D585FF 0%, #00FFEE 100%)"
+          >
+            A fantasy image manipulation relocating the habitat of wild animals.
+          </ProjectCard>
+        </ProjectsWrapper>
+      </Projects>
+      <About>
+        <Title>About</Title>
+        <AboutHero>
+          <Avatar src={avatar} alt="John Doe" />
+          <AboutSub>
+            The English language can not fully capture the depth and complexity of my thoughts. So I'm incorporating
+            Emoji into my speech to better express myself. Winky face.
+          </AboutSub>
+        </AboutHero>
+        <AboutDesc>
+          You know the way you feel when you see a picture of two otters holding hands? That's how you're gonna feel
+          every day. My mother cried the day I was born because she knew she’d never be prettier than me. You should
+          make me your campaign manager. I was born for politics. I have great hair and I love lying. Captain? The kids
+          want to know where Paulie the Pigeon is. I told them he got sucked up into an airplane engine, is that all
+          right?
+        </AboutDesc>
+      </About>
+      <Contact>
         <Inner>
           <Title>Get in touch</Title>
           <ContactText>
@@ -283,24 +128,10 @@ const Index = () => (
         </Inner>
         <Footer>
           &copy; 2018 by Gatsby Starter Portfolio Cara.{' '}
-          <a href="https://github.com/LekoArts/gatsby-starter-portfolio-cara">Github Repository</a>.
+          <a href="https://github.com/LekoArts/gatsby-starter-portfolio-cara">Github Repository</a>. Made by{' '}
+          <a href="https://www.lekoarts.de">LekoArts</a>.
         </Footer>
-      </Content>
-      <Divider speed={0.1} offset={4}>
-        <UpDown>
-          <SVG icon="upDown" hiddenMobile width={8} fill={colors['grey-darkest']} left="70%" top="20%" />
-          <SVG icon="triangle" width={8} stroke={colors['grey-darkest']} left="25%" top="5%" />
-        </UpDown>
-        <UpDownWide>
-          <SVG icon="triangle" width={12} stroke={colors.white} left="95%" top="50%" />
-          <SVG icon="circle" width={6} fill={colors.white} left="85%" top="15%" />
-          <SVG icon="upDown" hiddenMobile width={8} fill={colors['grey-darkest']} left="45%" top="10%" />
-        </UpDownWide>
-        <SVG icon="circle" width={6} fill={colors.white} left="4%" top="20%" />
-        <SVG icon="circle" width={12} fill={colors['grey-darkest']} left="70%" top="60%" />
-        <SVG icon="box" width={12} fill={colors['grey-darkest']} left="20%" top="30%" />
-        <SVG icon="hexa" width={8} stroke={colors['grey-darkest']} left="80%" top="70%" />
-      </Divider>
+      </Contact>
     </Parallax>
   </>
 )
